Rename ottoext to internal
