+++ geth/jail/internal/promise/promise.go
@@ -1,9 +1,9 @@
 package promise
 
 import (
-	"github.com/status-im/status-go/geth/jail/ottoext/loop"
-	"github.com/status-im/status-go/geth/jail/ottoext/timers"
-	"github.com/status-im/status-go/geth/jail/ottoext/vm"
+	"github.com/status-im/status-go/geth/jail/internal/loop"
+	"github.com/status-im/status-go/geth/jail/internal/timers"
+	"github.com/status-im/status-go/geth/jail/internal/vm"
 )
 
 func Define(vm *vm.VM, l *loop.Loop) error {
