+++ geth/jail/internal/timers/timers.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/robertkrimen/otto"
 
-	"github.com/status-im/status-go/geth/jail/ottoext/loop"
-	"github.com/status-im/status-go/geth/jail/ottoext/vm"
+	"github.com/status-im/status-go/geth/jail/internal/loop"
+	"github.com/status-im/status-go/geth/jail/internal/vm"
 )
 
 var minDelay = map[bool]int64{
