+++ geth/jail/jail_cell.go
@@ -2,10 +2,10 @@ package jail
 
 import (
 	"github.com/robertkrimen/otto"
-	"github.com/status-im/status-go/geth/jail/ottoext/fetch"
-	"github.com/status-im/status-go/geth/jail/ottoext/loop"
-	"github.com/status-im/status-go/geth/jail/ottoext/timers"
-	"github.com/status-im/status-go/geth/jail/ottoext/vm"
+	"github.com/status-im/status-go/geth/jail/internal/fetch"
+	"github.com/status-im/status-go/geth/jail/internal/loop"
+	"github.com/status-im/status-go/geth/jail/internal/timers"
+	"github.com/status-im/status-go/geth/jail/internal/vm"
 )
 
 // Cell represents a single jail cell, which is basically a JavaScript VM.
